Require `n::Integer` for most `sizehint!(a, n)`
Currently we're a bit hit-and-miss about any type-restriction on the
second argument to `sizehint!`: of 11 methods, 4 require `n::Integer`
and the remainder are unrestricted. Arguably, our dispatch is "prettier"
if methods differ only in their first argument and are consistent about
type-restrictions on `n`.

This imposes `::Integer` restrictions on all methods that specify a
non-abstract type for the first argument. Only the fallback methods for
`AbstractDict`, `AbstractSet`, and `AbstractArray` are left
unrestricted.

On 1.12-rc1, JET shows 40 possible errors for `report_opt(merge!,
(AbstractDict, AbstractDict))`, most of which are triggered from a
branch on `sizehint!(::WeakKeyDict, ::Any)`. This seems likely to fix
most of those errors.

+++ base/idset.jl
@@ -78,7 +78,7 @@ pop!(s::IdSet, @nospecialize(x)) = _pop!(s, x) == -1 ? throw(KeyError(x)) : x
 pop!(s::IdSet, @nospecialize(x), @nospecialize(default)) = _pop!(s, x) == -1 ? default : x
 delete!(s::IdSet, @nospecialize(x)) = (_pop!(s, x); s)
 
-function sizehint!(s::IdSet, newsz)
+function sizehint!(s::IdSet, newsz::Integer)
     # TODO: grow/compact list and perform rehash, if profitable?
     # TODO: shrink?
     # s.list = resize(s.list, newsz)
